Resources + critical bug fixed update
1) Removed global theme resources.  To update local theme resources better use AddThemeResourcesNotifier.
2) Critical bug fixed for Setter.
3) Added OpenUrl.
And other small changes.

+++ Sgl/Base/Ref.h
@@ -11,7 +11,7 @@ namespace Sgl
     {
         virtual ~RefMemoryBlockBase() = default;
 
-        std::atomic<size_t> References = 1;
+        std::atomic<uint32_t> References = 1;
     };
 
     template<typename T>
